Find packages for meshes / extras urdf (#45)
* Allow meshes visual and extras urdf to be linked using find package functionality
